+++ lib/ts/recipe/session/types.ts
@@ -67,6 +67,7 @@ export type TypeInput = {
     cookieSecure?: boolean;
     cookieSameSite?: "strict" | "lax" | "none";
     cookieDomain?: string;
+    olderCookieDomain?: string;
     overwriteSessionDuringSignInUp?: boolean;
 
     getTokenTransferMethod?: (input: {
@@ -112,6 +113,7 @@ export type TypeNormalisedInput = {
     refreshTokenPath: NormalisedURLPath;
     accessTokenPath: NormalisedURLPath;
     cookieDomain: string | undefined;
+    olderCookieDomain: string | undefined;
     getCookieSameSite: (input: {
         request: BaseRequest | undefined;
         userContext: UserContext;
@@ -410,7 +412,10 @@ export type APIInterface = {
      */
     refreshPOST:
         | undefined
-        | ((input: { options: APIOptions; userContext: UserContext }) => Promise<SessionContainerInterface>);
+        | ((input: {
+              options: APIOptions;
+              userContext: UserContext;
+          }) => Promise<SessionContainerInterface | undefined>);
 
     signOutPOST:
         | undefined